requirements.txt
appdirs==1.4.3
Django==2.1.7
djangorestframework==3.9.2
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
psycopg2==2.7.3.2
gunicorn==19.6.0

